Update rest model
