+++ src/main/java/fhirspark/restmodel/Modification.java
@@ -1,11 +1,9 @@
 
 package fhirspark.restmodel;
 
-import java.io.Serializable;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import org.apache.commons.lang3.builder.ToStringBuilder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
@@ -13,8 +11,7 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
     "recommender",
     "timestamp"
 })
-public class Modification implements Serializable
-{
+public class Modification {
 
     @JsonProperty("modified")
     private String modified;
@@ -22,7 +19,6 @@ public class Modification implements Serializable
     private Recommender recommender;
     @JsonProperty("timestamp")
     private String timestamp;
-    private final static long serialVersionUID = 7081567173278031069L;
 
     @JsonProperty("modified")
     public String getModified() {
@@ -52,11 +48,6 @@ public class Modification implements Serializable
     @JsonProperty("timestamp")
     public void setTimestamp(String timestamp) {
         this.timestamp = timestamp;
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this).append("modified", modified).append("recommender", recommender).append("timestamp", timestamp).toString();
     }
 
 }
