+++ src/main/java/fhirspark/restmodel/TherapyRecommendation.java
@@ -1,26 +1,57 @@
 
 package fhirspark.restmodel;
 
-import java.io.Serializable;
 import java.util.List;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import org.apache.commons.lang3.builder.ToStringBuilder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
+    "comment",
+    "evidenceLevel",
     "id",
-    "therapyRecommendations"
+    "modifications",
+    "reasoning",
+    "references",
+    "treatments"
 })
-public class TherapyRecommendation implements Serializable
-{
+public class TherapyRecommendation {
 
+    @JsonProperty("comment")
+    private List<String> comment = null;
+    @JsonProperty("evidenceLevel")
+    private String evidenceLevel;
     @JsonProperty("id")
     private String id;
-    @JsonProperty("therapyRecommendations")
-    private List<TherapyRecommendation_> therapyRecommendations = null;
-    private final static long serialVersionUID = 8346910927803718627L;
+    @JsonProperty("modifications")
+    private List<Modification> modifications = null;
+    @JsonProperty("reasoning")
+    private Reasoning reasoning;
+    @JsonProperty("references")
+    private List<Reference> references = null;
+    @JsonProperty("treatments")
+    private List<Treatment> treatments = null;
+
+    @JsonProperty("comment")
+    public List<String> getComment() {
+        return comment;
+    }
+
+    @JsonProperty("comment")
+    public void setComment(List<String> comment) {
+        this.comment = comment;
+    }
+
+    @JsonProperty("evidenceLevel")
+    public String getEvidenceLevel() {
+        return evidenceLevel;
+    }
+
+    @JsonProperty("evidenceLevel")
+    public void setEvidenceLevel(String evidenceLevel) {
+        this.evidenceLevel = evidenceLevel;
+    }
 
     @JsonProperty("id")
     public String getId() {
@@ -32,19 +63,44 @@ public class TherapyRecommendation implements Serializable
         this.id = id;
     }
 
-    @JsonProperty("therapyRecommendations")
-    public List<TherapyRecommendation_> getTherapyRecommendations() {
-        return therapyRecommendations;
+    @JsonProperty("modifications")
+    public List<Modification> getModifications() {
+        return modifications;
     }
 
-    @JsonProperty("therapyRecommendations")
-    public void setTherapyRecommendations(List<TherapyRecommendation_> therapyRecommendations) {
-        this.therapyRecommendations = therapyRecommendations;
+    @JsonProperty("modifications")
+    public void setModifications(List<Modification> modifications) {
+        this.modifications = modifications;
     }
 
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this).append("id", id).append("therapyRecommendations", therapyRecommendations).toString();
+    @JsonProperty("reasoning")
+    public Reasoning getReasoning() {
+        return reasoning;
+    }
+
+    @JsonProperty("reasoning")
+    public void setReasoning(Reasoning reasoning) {
+        this.reasoning = reasoning;
+    }
+
+    @JsonProperty("references")
+    public List<Reference> getReferences() {
+        return references;
+    }
+
+    @JsonProperty("references")
+    public void setReferences(List<Reference> references) {
+        this.references = references;
+    }
+
+    @JsonProperty("treatments")
+    public List<Treatment> getTreatments() {
+        return treatments;
+    }
+
+    @JsonProperty("treatments")
+    public void setTreatments(List<Treatment> treatments) {
+        this.treatments = treatments;
     }
 
 }
