+++ src/main/java/fhirspark/restmodel/Treatment.java
@@ -1,11 +1,9 @@
 
 package fhirspark.restmodel;
 
-import java.io.Serializable;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import org.apache.commons.lang3.builder.ToStringBuilder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
@@ -13,8 +11,7 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
     "ncit_code",
     "synonyms"
 })
-public class Treatment implements Serializable
-{
+public class Treatment {
 
     @JsonProperty("name")
     private String name;
@@ -22,7 +19,6 @@ public class Treatment implements Serializable
     private String ncitCode;
     @JsonProperty("synonyms")
     private String synonyms;
-    private final static long serialVersionUID = 5378130599841238154L;
 
     @JsonProperty("name")
     public String getName() {
@@ -52,11 +48,6 @@ public class Treatment implements Serializable
     @JsonProperty("synonyms")
     public void setSynonyms(String synonyms) {
         this.synonyms = synonyms;
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this).append("name", name).append("ncitCode", ncitCode).append("synonyms", synonyms).toString();
     }
 
 }
